Moved KNOWN_LINKS into tools.jsx from AssistantRuleBook.jsx
