+++ src/components/NavItems/tools/Archive/components/urlArchive.jsx
@@ -11,13 +11,13 @@ import Typography from "@mui/material/Typography";
 
 import { useTrackEvent } from "@/Hooks/useAnalytics";
 import { ROLES } from "@/constants/roles";
+import { KNOWN_LINKS } from "@/constants/tools";
 import { getclientId } from "@Shared/GoogleAnalytics/MatomoAnalytics";
 import { history } from "@Shared/History/History";
 import { i18nLoadNamespace } from "components/Shared/Languages/i18nLoadNamespace";
 
 import IconInternetArchive from "../../../../NavBar/images/SVG/Others/archive-icon.svg";
 import CopyButton from "../../../../Shared/CopyButton";
-import { KNOWN_LINKS } from "../../../Assistant/AssistantRuleBook";
 import { prettifyLargeString } from "../utils";
 import DownloadWaczFile from "./downloadWaczFile";
